% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_gower}
\alias{calc_gower}
\title{Calculate Gower dissimilarity (two vectors)}
\usage{
calc_gower(vec_from, vec_to = NULL)
}
\arguments{
\item{vec_from}{Numeric vector (site A).}

\item{vec_to}{Numeric vector (site B). If \code{NULL}, returns \code{NA_real_}.}
}
\value{
Numeric scalar in the range 0 to 1 or \code{NA_real_}.
}
\description{
Uses \code{cluster::daisy()} with metric = "gower".
}
\examples{
calc_gower(c(3, 0, 1, 2), c(0, 4, 1, 5))
calc_gower(c(1, 2, 3), NULL)
}
\keyword{internal}
