% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissmapr-package.R
\docType{package}
\name{dissmapr-package}
\alias{dissmapr}
\alias{dissmapr-package}
\title{dissmapr: Workflow for Compositional Dissimilarity and Biodiversity Turnover Analysis}
\description{
Tools to compute and map biodiversity turnover at the macroscale. Provides functions for generating spatial grids, retrieving environmental data, computing multi-order zeta diversity, fitting I-spline models for generalised dissimilarity, predicting community turnover, and delineating and mapping bioregions and their changes over time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/b-cubed-eu/dissmapr}
  \item \url{https://b-cubed-eu.github.io/dissmapr/}
  \item Report bugs at \url{https://github.com/b-cubed-eu/dissmapr/issues}
}

}
\author{
\strong{Maintainer}: Sandra MacFadyen \email{macfadyen@sun.ac.za} (\href{https://orcid.org/0000-0002-5316-440X}{ORCID})

Authors:
\itemize{
  \item Sandra MacFadyen \email{macfadyen@sun.ac.za} (\href{https://orcid.org/0000-0002-5316-440X}{ORCID})
}

Other contributors:
\itemize{
  \item Stellenbosch University [copyright holder]
}

}
\keyword{internal}
